Move cgroup resolver to its own package
